The heap is not capped at 4 GB: addresses are 48-bit everywhere
`Value::heap(off: u32)` capped flint's heap at 4 GB on every runtime. That
was an artificial limit inherited from wasm32, not a property of the
representation: `TAG_HEAP` occupies bits 63..48 and leaves 48 payload bits
-- 256 TB, or 2 PB if the eight-byte alignment were folded in.

`mem::Addr` is now that width, and it is `u64` EVERYWHERE rather than
per-target. It costs nothing on wasm32, where an address simply never
exceeds 32 bits, and it buys two things a cfg would not: wasm64 works with
no change and no configuration, and there is ONE width to reason about, so
a snapshot, an image and a `Value` mean the same thing on every runtime
rather than nearly the same thing.

TWO PLACES STORED AN ADDRESS IN 32 BITS, and the compiler found both:

* A FORWARDING POINTER lived in the object's `len` field. It needs no
  wider header: word 0 is `[31:24] type [23:21] age [20] mark [19] remset`
  and a `TY_FWD` has no age, no mark and is in no remembered set, so its
  low 24 bits are free. 24 + 32 = 56 against the 48 an address can be.
  `set_forward`/`forward_target` live together so the packing cannot
  drift -- splitting them is how a reader ends up with half an address.

* The FREE LIST threaded its next-pointer through `write_u32(addr + 8)`.
  The `size < 16` guard above it is what reserves the room: a linked block
  always has those eight bytes free, which is exactly a wide pointer.

WHAT DID NOT WIDEN, and the type is what says so. `native.rs` reads wasm
DATA SEGMENTS -- linear-memory addresses, 32-bit by the platform's
definition and nothing to do with flint's heap. The wasm arena is the
same: it backs Rust's own allocator below the flint heap, in memory that
cannot exceed 4 GB. Both stay `u32`, and `Space::take` is now the one
place the two address spaces meet.

The diagnostic `stat_*` exports still hand the host a `u32`, so `narrow`
does that in one place and says plainly that it truncates above 4 GB and
why that is tolerated only there: they are bisection aids, absent from a
production build entirely.

Snapshot format 2 -> 3, and the inspector reads addresses as `Number` --
48 bits is inside the 53 a double holds exactly, so an address is never
approximated.

`heap_refs` now round-trips the WHOLE payload, including
`0x0000_FFFF_FFFF_FFF8`, rather than stopping at `u32::MAX`.

322 compile errors at the widest point. Suite green, 10/10 host
conformance, both build configurations, all three unit builds.

NOT MEASURED: what this costs on wasm. The 32-bit and 48-bit modules came
out byte-identical, which would mean LLVM narrows it all back -- but my
control for "does the build respond to a change at all" did not produce a
file, so that is an unverified coincidence rather than a result. Saying so
rather than reporting a number I cannot stand behind.

+++ runtime/src/vm.rs
@@ -1601,7 +1601,7 @@ impl Rt {
         let limit = self.gc.heap_limit();
         if self.mem_trips == 0 {
             self.mem_trips = 1;
-            self.gc.set_heap_limit(limit.saturating_add(1024 * 1024));
+            self.gc.set_heap_limit(limit.saturating_add(1024 * 1024) as u32);
         }
         let used = self.gc.heap_used();
         let msg = alloc::format!(
